!68 修订部分问题, 清理package * 删除早期的测试代码 * 删除早期的测试代码 * 删除tdx和v1版本的dataframe * 删除data包, 引入gitee.com/quant1x/data工具包 * 更新gox版本 * 更新gitee脚本, 只做pull相关操作, 方便在github脚本中同步 * 调整部分函数的判断逻辑 * series 补充逻辑判断的方法 * 补充新扩展的接口, 暂未实现 * 修订注释, 增加测试代码 * 修复浮点0.xx会认为是字符串的bug * 公式返回值序列化调整 * 删除无用的代码 * 删除无用的代码 * 删除无用的代码 * 删除无用的代码 * 修定参数名 * 修定参数名 * 修正注释 * 修复序列版本不支持负值向前移动 * 修复序列版本不支持负值向前移动 * 修正对齐函数, 长度相等直接返回 * 修订兼容性

+++ stat/series_convert.go
@@ -64,7 +64,7 @@ func ToBool(s Series) []bool {
 	case SERIES_TYPE_FLOAT32:
 		return __NumberToBool_S(values.([]float32))
 	case SERIES_TYPE_FLOAT64:
-		return __NumberToBool_S(values.([]int64))
+		return __NumberToBool_S(values.([]float64))
 	case SERIES_TYPE_INT32:
 		return __NumberToBool_S(values.([]int32))
 	case SERIES_TYPE_INT64:
